Ditch redo, just use a shell script.
We don't need a fancy build system right now: the project is so small
that partial compilation isn't a big win, and it remains to be seen
how tangled the web of dependencies will be. Given that and redo's
portability problems, it's best to hold off as long as possible.